refactor: extract CLI recipe source helpers
